build(ControllerMissingNamespace): Delete this policy altogether
This policy checks that operation requests on namespaced resources have
a metadata.namespace set.

Yet the Kubernetes API would not accept a namespaced resource without a
namespace specified in the admission request that the API gets. It
would, at minimum, add the defined default namespace.

The Kubernetes apiserver already has an admission controller for this case even:
https://kubernetes.io/docs/reference/access-authn-authz/admission-controllers/#namespaceexists

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>